Refined

Description

This is the main entry point to the library. It provides the core types and functions for usage.

Types

Refined

data Refined ps a where

This is the core type for Refined, a newtype over a that has a type-level list of attached predicates. The list should be thought of as a conjunction (AND) of clauses, where each clause is a disjunction (OR) of literals. In other words, the list is in conjunctive normal form, with the caveat that we also allow exclusive disjunction (XOR).

For example,

  Refined '[NonZero, Positive, Xor Even Odd] Int
  Refined '[MaxLength 100, Not Upper, Or Alpha Numeric] Text

are in CNF, while

  Refined '[Not (Or Alpha Numeric)] Text

is not.

Since: 0.1.0.0

Bundled Patterns

pattern MkRefined :: a -> Refined p a
Since: 0.1.0.0

refineEmpty :: a -> Refined '[] a

Wraps a in Refined with no attached predicates.

Examples

Expand

>>> refineEmpty 0
UnsafeRefined {unrefine = 0}

Since: 0.1.0.0

refine :: forall p a. Predicate p a => a -> Either RefineException (Refined '[p] a)

Attempts to prove the given predicate. If it succeeds, we return the refined a. Otherwise we return an error.

Examples

Expand

>>> let x = refine @Positive @Int 5
>>> :type x
x :: Either RefineException (Refined '[Positive] Int)

>>> x
Right (UnsafeRefined {unrefine = 5})

>>> refine @Positive 0
Left (MkRefineException {predRep = GreaterThan 0, targetRep = Integer, msg = "0 does not satisfy > 0"})

Since: 0.1.0.0

refineAll :: forall ps a. Predicate (Proxy ps) a => a -> Either RefineException (Refined ps a)

Attempts to prove multiple predicates. Returns the first error encountered, if any.

Examples

Expand

>>> let x = refineAll @'[Positive, Odd, (GreaterThan 2)] @Int 5
>>> :type x
x :: Either
       RefineException (Refined '[Positive, Odd, GreaterThan 2] Int)

>>> x
Right (UnsafeRefined {unrefine = 5})

>>> refineAll @'[Positive, Odd, GreaterThan 6] @Int 5
Left (MkRefineException {predRep = GreaterThan 6, targetRep = Int, msg = "5 does not satisfy > 6"})

Since: 0.1.0.0

refineTH :: forall p a. (Predicate p a, Lift a) => a -> Q (TExp (Refined '[p] a))

Proves p at compile-time via TemplateHaskell.

Examples

Expand

>>> let x = $$((refineTH @NonZero 5))
>>> :type x
x :: Refined '[NonZero] Integer

>>> x
UnsafeRefined {unrefine = 5}

Since: 0.1.0.0

refineAllTH :: forall ps a. (Predicate (Proxy ps) a, Lift a) => a -> Q (TExp (Refined ps a))

Proves multiple predicates at compile-time via TemplateHaskell. See refineAll.

Examples

Expand

>>> let x = $$((refineAllTH @[NonZero, Odd, Positive] 5))
>>> :type x
x :: Refined '[NonZero, Odd, Positive] Integer

>>> x
UnsafeRefined {unrefine = 5}

Since: 0.1.0.0

unsafeRefine :: forall p a. Predicate p a => a -> Refined '[p] a

Attempts to prove the given predicate. If it succeeds, we return the refined a. Otherwise we die with a runtime error.

>>> unsafeRefine @NonNegative 0
UnsafeRefined {unrefine = 0}

Since: 0.1.0.0

unsafeRefineAll :: forall ps a. Predicate (Proxy ps) a => a -> Refined ps a

Attempts to prove the given predicates. If it succeeds, we return the refined a. Otherwise we die with a runtime error.

>>> unsafeRefineAll @'[NonNegative, Even] 4
UnsafeRefined {unrefine = 4}

Since: 0.1.0.0

RefineException

data RefineException

RefineException is used for errors encountered during refinement.

Since: 0.1.0.0

Constructors

MkRefineException

Fields

predRep :: TypeRep

Type rep for the predicate we tried to apply.

Since: 0.1.0.0

targetRep :: TypeRep

Type rep for the type to which we tried to apply the refinement.

Since: 0.1.0.0

msg :: String

Description of the error encountered.

Since: 0.1.0.0

mkRefineException :: forall p a. (Predicate p a, Typeable p, Typeable a) => String -> RefineException

mkRefineException @p @a msg creates a RefineException for type a, predicate p, and error message msg.

Since: 0.1.0.0

Predicates

class Predicate p a where

This class is used for defining new predicates.

We include instances for type-level lists so that we can prove multiple predicates simultaneously.

Since: 0.1.0.0

Methods

satisfies :: Proxy p -> a -> Maybe RefineException

Validates predicate p for the type a. Returns Nothing on success, Just RefineException on a validation failure.

Since: 0.1.0.0

Operators

data Not p

Logical negation.

Examples

Expand

>>> satisfies (Proxy @(Not Positive)) (-2)
Nothing

>>> satisfies (Proxy @(Not Positive)) 7
Just (MkRefineException {predRep = Not (GreaterThan 0), targetRep = Integer, msg = "7 does not satisfy (Not (GreaterThan 0))"})

Since: 0.1.0.0

data Or p q

Logical disjunction.

Examples

Expand

>>> satisfies @(Or NonZero Positive) Proxy (-2)
Nothing

>>> satisfies @(Or NonZero Positive) Proxy 7
Nothing

>>> satisfies @(Or NonZero Positive) Proxy 0
Just (MkRefineException {predRep = Or (Not (NatEquals 0)) (GreaterThan 0), targetRep = Integer, msg = "0 satisfied neither predicate"})

Since: 0.1.0.0

type (\/) = Or infixr 2

Infix operator for Or.

Since: 0.1.0.0

data Xor p q

Logical exclusive disjunction.

Examples

Expand

>>> satisfies @(Xor NonZero Positive) Proxy (-2)
Nothing

>>> satisfies @(Xor NonZero Positive) Proxy 7
Just (MkRefineException {predRep = Xor (Not (NatEquals 0)) (GreaterThan 0), targetRep = Integer, msg = "7 satisfied both predicates"})

>>> satisfies @(Xor NonZero Positive) Proxy 0
Just (MkRefineException {predRep = Xor (Not (NatEquals 0)) (GreaterThan 0), targetRep = Integer, msg = "0 satisfied neither predicate"})

Since: 0.1.0.0

type (<+>) = Xor infixr 3

Infix operator for Xor.

Since: 0.1.0.0

Proving

addPred :: forall p ps a. Predicate p a => Refined ps a -> Either RefineException (Refined (AppendP p ps) a)

Attempts to prove the given predicate. If it succeeds, we add the predicate to the list.

Examples

Expand

>>> let x = unsafeRefine @NonNegative @Int 7
>>> y = addPred @NonZero x
>>> :type y
y :: Either
       RefineException (Refined '[GreaterThanEq 0, Not (NatEquals 0)] Int)

>>> y
Right (UnsafeRefined {unrefine = 7})

>>> let z = addPred @Even x
>>> z
Left (MkRefineException {predRep = Even, targetRep = Int, msg = "7 is not even"})

Since: 0.1.0.0

unsafeAddPred :: forall p ps a. Predicate p a => Refined ps a -> Refined (AppendP p ps) a

Attempts to prove the given predicate. If it succeeds, we add the predicate to the list.

Examples

Expand

>>> let x = unsafeRefine @NonNegative @Int 7
>>> y = unsafeAddPred @NonZero x
>>> :type y
y :: Refined '[GreaterThanEq 0, Not (NatEquals 0)] Int

>>> y
UnsafeRefined {unrefine = 7}

Since: 0.1.0.0

relax :: forall p ps a. Refined ps a -> Refined (DeleteP p ps) a

Removes all occurrences of the desired predicate, if it exists.

Examples

Expand

>>> let x = unsafeRefine @Even @Int 8
>>> x' = relax @Even x
>>> :type x'
x' :: Refined '[] Int

Since: 0.1.0.0

relaxAll :: Refined ps a -> Refined '[] a

Removes all predicates.

Examples

Expand

>>> let x = unsafeRefine @Even @Int 8
>>> y = unsafeAddPred @Positive x
>>> :type y
y :: Refined '[Even, GreaterThan 0] Int

>>> :type (relaxAll y)
(relaxAll y) :: Refined '[] Int

Since: 0.1.0.0

TypeErrors

type family Implies ps p where ...

Implies ps p raises a type error if ps does not logically imply p. See ImpliesBoolExpr for more information on when ps implies p.

Examples

Expand

>>> -- safeDiv :: (Implies ps NonZero, Integral n) => n -> Refined ps n -> n
>>> let d = unsafeRefine @NonZero @Int 7
>>> safeDiv 14 d
2

Since: 0.1.0.0

Equations

Implies ps p = ErrIfFalse (PredNotFound ps p) (ImpliesBool ps p)

type PredNotFound ps p = (('Text "Desired predicate " :<>: 'ShowType p) :<>: 'Text " was not found in list ") :<>: 'ShowType ps

Type error for a missing, wanted predicate.

Since: 0.1.0.0

Built-In

module Refined.Predicate.Math

module Refined.Predicate.Foldable

module Refined.Predicate.Text

Misc Utils

type family AppendP p ps where ...

AppendP p ps appends p to ps if it does not already occur in the list.

Since: 0.1.0.0

Equations

AppendP p '[] = '[p]

AppendP p (p ': ps) = p ': ps

AppendP p (q ': ps) = q ': AppendP p ps

type family DeleteP p ps where ...

DeleteP p ps deletes all occurrences of p from ps.

Since: 0.1.0.0

Equations

DeleteP _ '[] = '[]

DeleteP p (p ': ps) = DeleteP p ps

DeleteP p (q ': ps) = q ': DeleteP p ps

type family ImpliesBool ps p where ...

Variant of Implies that returns a Bool.

Since: 0.1.0.0

Equations

ImpliesBool '[] _ = 'False

ImpliesBool (p ': _) p = 'True

ImpliesBool (q ': ps) p = ImpliesBoolExpr q p || ImpliesBool ps p

type family ImpliesBoolExpr expr p where ...

ImpliesBoolExpr expr p returns 'True if expr logically implies p. expr is assumed to be a combination of predicate literals (i.e. a predicate with no logical connectives) with logical combinators from Operators.

NB. For a (possibly exclusive) disjunction to imply p, both clauses must individually imply p, since we cannot know which one is satisfied. That is,

\[ q_1 \vee q_2 \implies p \quad \iff \quad (q_1 \implies p) \wedge (q_2 \implies p) \]

Examples

Expand

>>> :kind! ImpliesBoolExpr (Not NonZero) (NonNegative)
ImpliesBoolExpr (Not NonZero) (NonNegative) :: Bool
= 'False

>>> :kind! ImpliesBoolExpr (Not (Not NonNegative)) NonNegative
ImpliesBoolExpr (Not (Not NonNegative)) NonNegative :: Bool
= 'True

>>> :kind! ImpliesBoolExpr (NonZero \/ Alpha) NonZero
ImpliesBoolExpr (NonZero \/ Alpha) NonZero :: Bool
= 'False

>>> :kind! ImpliesBoolExpr (Alpha \/ (Not (Not Alpha))) Alpha
ImpliesBoolExpr (Alpha \/ (Not (Not Alpha))) Alpha :: Bool
= 'True

Since: 0.1.0.0

Equations

ImpliesBoolExpr p p = 'True

ImpliesBoolExpr (Not p) p = 'False

ImpliesBoolExpr (Not expr) p = ImpliesBoolExpr expr (Not p)

ImpliesBoolExpr (e1 \/ e2) p = ImpliesBoolExpr e1 p && ImpliesBoolExpr e2 p

ImpliesBoolExpr (e1 <+> e2) p = ImpliesBoolExpr e1 p && ImpliesBoolExpr e2 p

ImpliesBoolExpr q p = 'False

type family ErrIfFalse p b where ...

Emits a TypeError when given False. The parameter type is used in the error message.

Since: 0.1.0.0

Equations

ErrIfFalse err 'False = TypeError err

ErrIfFalse _ _ = ()