Refined.Internal

Description

Internal module that exports types, including the unsafe UnsafeRefined.

Since: 0.1.0.0

Documentation

newtype Refined ps a

This is the core type for Refined, a newtype over a that has a type-level list of attached predicates. The list should be thought of as a conjunction (AND) of clauses, where each clause is a disjunction (OR) of literals. In other words, the list is in conjunctive normal form, with the caveat that we also allow exclusive disjunction (XOR).

For example,

  Refined '[NonZero, Positive, Xor Even Odd] Int
  Refined '[MaxLength 100, Not Upper, Or Alpha Numeric] Text

are in CNF, while

  Refined '[Not (Or Alpha Numeric)] Text

is not.

Since: 0.1.0.0

Constructors

UnsafeRefined

Fields

unrefine :: a

Since: 0.1.0.0

Bundled Patterns

pattern MkRefined :: a -> Refined p a
Since: 0.1.0.0

data RefineException

RefineException is used for errors encountered during refinement.

Since: 0.1.0.0

Constructors

MkRefineException

Fields

predRep :: TypeRep

Type rep for the predicate we tried to apply.

Since: 0.1.0.0

targetRep :: TypeRep

Type rep for the type to which we tried to apply the refinement.

Since: 0.1.0.0

msg :: String

Description of the error encountered.

Since: 0.1.0.0