Refined.Predicate

Description

This module provides the Predicate typeclass along with some built-in predicates.

Since: 0.1.0.0

Typeclass

class Predicate p a where

This class is used for defining new predicates.

We include instances for type-level lists so that we can prove multiple predicates simultaneously.

Since: 0.1.0.0

Methods

satisfies :: Proxy p -> a -> Maybe RefineException

Validates predicate p for the type a. Returns Nothing on success, Just RefineException on a validation failure.

Since: 0.1.0.0

mkRefineException :: forall p a. (Predicate p a, Typeable p, Typeable a) => String -> RefineException

mkRefineException @p @a msg creates a RefineException for type a, predicate p, and error message msg.

Since: 0.1.0.0

Operators

data Not p

Logical negation.

Examples

Expand

>>> satisfies (Proxy @(Not Positive)) (-2)
Nothing

>>> satisfies (Proxy @(Not Positive)) 7
Just (MkRefineException {predRep = Not (GreaterThan 0), targetRep = Integer, msg = "7 does not satisfy (Not (GreaterThan 0))"})

Since: 0.1.0.0

data Or p q

Logical disjunction.

Examples

Expand

>>> satisfies @(Or NonZero Positive) Proxy (-2)
Nothing

>>> satisfies @(Or NonZero Positive) Proxy 7
Nothing

>>> satisfies @(Or NonZero Positive) Proxy 0
Just (MkRefineException {predRep = Or (Not (NatEquals 0)) (GreaterThan 0), targetRep = Integer, msg = "0 satisfied neither predicate"})

Since: 0.1.0.0

type (\/) = Or infixr 2

Infix operator for Or.

Since: 0.1.0.0

data Xor p q

Logical exclusive disjunction.

Examples

Expand

>>> satisfies @(Xor NonZero Positive) Proxy (-2)
Nothing

>>> satisfies @(Xor NonZero Positive) Proxy 7
Just (MkRefineException {predRep = Xor (Not (NatEquals 0)) (GreaterThan 0), targetRep = Integer, msg = "7 satisfied both predicates"})

>>> satisfies @(Xor NonZero Positive) Proxy 0
Just (MkRefineException {predRep = Xor (Not (NatEquals 0)) (GreaterThan 0), targetRep = Integer, msg = "0 satisfied neither predicate"})

Since: 0.1.0.0

type (<+>) = Xor infixr 3

Infix operator for Xor.

Since: 0.1.0.0

Built-in Predicates

module Refined.Predicate.Math

module Refined.Predicate.Foldable

module Refined.Predicate.Text