Refined.Predicate.Text

Description

Provides predicates for Text and String.

Since: 0.1.0.0

Symbol Equality

data SymEquals c

Predicate equality for symbols.

Examples

Expand

>>> satisfies @(SymEquals "c") Proxy 'c'
Nothing

>>> satisfies @(SymEquals "abc") Proxy 'c'
Just (MkRefineException {predRep = SymEquals "abc", targetRep = Char, msg = "c is not a single Char"})

>>> satisfies @(SymEquals "abc") Proxy "abc"
Nothing

>>> satisfies @(SymEquals "123") @Text Proxy "abc"
Just (MkRefineException {predRep = SymEquals "123", targetRep = Text, msg = "abc does not equal the predicate"})

Since: 0.1.0.0

type SymNotEquals c = Not (SymEquals c)

The logical negation of SymEquals.

Examples

Expand

>>> satisfies @(SymNotEquals "c") Proxy 'x'
Nothing

>>> satisfies @(SymNotEquals "abc") Proxy "abc"
Just (MkRefineException {predRep = Not (SymEquals "abc"), targetRep = [Char], msg = "\"abc\" does not satisfy (Not (SymEquals \"abc\"))"})

Since: 0.1.0.0

Char predicates

These section models the boolean functions defined in Data.Char. See that module for more information regarding these definitions.

Although the instances are defined for Char, they can be extended to String and Text via Refined.Predicate.Foldable's All.

Examples

Expand

>>> satisfies @(All Alpha) Proxy "abc"
Nothing

>>> satisfies @(All Alpha) Proxy "abc1"
Just (MkRefineException {predRep = Alpha, targetRep = Char, msg = "1 is not an alphabetic character"})

Character Classification

data Control

Predicate for isControl.

Examples

Expand

>>> satisfies @Control Proxy '\r'
Nothing

>>> satisfies @Control Proxy 'a'
Just (MkRefineException {predRep = Control, targetRep = Char, msg = "a is not a control character"})

Since: 0.1.0.0

data Space

Predicate for a isSpace.

Examples

Expand

>>> satisfies @Space Proxy ' '
Nothing
>>> satisfies @Space Proxy '\r'
Nothing

>>> satisfies @Space Proxy 'a'
Just (MkRefineException {predRep = Space, targetRep = Char, msg = "a is not a space character"})

Since: 0.1.0.0

data Lower

Predicate for isLower.

Examples

Expand

>>> satisfies @Lower Proxy 'c'
Nothing

>>> satisfies @Lower Proxy 'C'
Just (MkRefineException {predRep = Lower, targetRep = Char, msg = "C is not lowercase"})

Since: 0.1.0.0

data Upper

Predicate for isUpper.

Examples

Expand

>>> satisfies @Upper Proxy 'C'
Nothing

>>> satisfies @Upper Proxy 'c'
Just (MkRefineException {predRep = Upper, targetRep = Char, msg = "c is not uppercase"})

Since: 0.1.0.0

data Alpha

Predicate for isAlpha.

Examples

Expand

>>> satisfies @Alpha Proxy 'c'
Nothing

>>> satisfies @Alpha Proxy '7'
Just (MkRefineException {predRep = Alpha, targetRep = Char, msg = "7 is not an alphabetic character"})

Since: 0.1.0.0

data AlphaNum

Predicate for isAlphaNum.

Examples

Expand

>>> satisfies @AlphaNum Proxy 'a'
Nothing

>>> satisfies @AlphaNum Proxy '1'
Nothing

>>> satisfies @AlphaNum Proxy '!'
Just (MkRefineException {predRep = AlphaNum, targetRep = Char, msg = "! is not an alpha-numeric character"})

Since: 0.1.0.0

data Digit

Predicate for isDigit.

Examples

Expand

>>> satisfies @Digit Proxy '1'
Nothing

>>> satisfies @Digit Proxy 'a'
Just (MkRefineException {predRep = Digit, targetRep = Char, msg = "a is not a digit"})

Since: 0.1.0.0

data OctDigit

Predicate for isOctDigit.

Examples

Expand

>>> satisfies @OctDigit Proxy '4'
Nothing

>>> satisfies @OctDigit Proxy '9'
Just (MkRefineException {predRep = OctDigit, targetRep = Char, msg = "9 is not an octal digit"})

Since: 0.1.0.0

data HexDigit

Predicate for isHexDigit.

>>> satisfies @HexDigit Proxy '1'
Nothing

Examples

Expand

>>> satisfies @HexDigit Proxy 'f'
Nothing

>>> satisfies @HexDigit Proxy 'g'
Just (MkRefineException {predRep = HexDigit, targetRep = Char, msg = "g is not a hexadecimal digit"})

Since: 0.1.0.0

data Letter

Predicate for isLetter.

Examples

Expand

>>> satisfies @Letter Proxy 'f'
Nothing

>>> satisfies @Letter Proxy '\r'
Just (MkRefineException {predRep = Letter, targetRep = Char, msg = "\r is not a letter"})

Since: 0.1.0.0

data Mark

Predicate for isMark.

Examples

Expand

>>> satisfies @Mark Proxy '\x20DD'
Nothing

>>> satisfies @Mark Proxy 'a'
Just (MkRefineException {predRep = Mark, targetRep = Char, msg = "a is not a mark"})

Since: 0.1.0.0

data Number

Predicate for isNumber.

Examples

Expand

>>> satisfies @Number Proxy '2'
Nothing

>>> satisfies @Number Proxy 'a'
Just (MkRefineException {predRep = Number, targetRep = Char, msg = "a is not a number"})

Since: 0.1.0.0

data Punctuation

Predicate for isPunctuation.

Examples

Expand

>>> satisfies @Punctuation Proxy '!'
Nothing

>>> satisfies @Punctuation Proxy 'a'
Just (MkRefineException {predRep = Punctuation, targetRep = Char, msg = "a is not punctuation"})

Since: 0.1.0.0

data Symbol

Predicate for isSymbol.

Examples

Expand

>>> satisfies @Symbol Proxy '$'
Nothing

>>> satisfies @Symbol Proxy 'a'
Just (MkRefineException {predRep = Symbol, targetRep = Char, msg = "a is not a symbol"})

Since: 0.1.0.0

data Separator

Predicate for isSeparator.

Examples

Expand

>>> satisfies @Separator Proxy ' '
Nothing

>>> satisfies @Separator Proxy 'a'
Just (MkRefineException {predRep = Separator, targetRep = Char, msg = "a is not a separator"})

Since: 0.1.0.0

Subranges

data Ascii

Predicate for isAscii.

Examples

Expand

>>> satisfies @Ascii Proxy 'a'
Nothing

>>> satisfies @Ascii Proxy '\x20DD'
Just (MkRefineException {predRep = Ascii, targetRep = Char, msg = "\8413 is not ascii"})

Since: 0.1.0.0

data Latin1

Predicate for Latin1.

Examples

Expand

>>> satisfies @Latin1 Proxy 'a'
Nothing

>>> satisfies @Latin1 Proxy '\x20DD'
Just (MkRefineException {predRep = Latin1, targetRep = Char, msg = "\8413 is not latin1"})

Since: 0.1.0.0

data AsciiUpper

Predicate for isAsciiUpper.

Examples

Expand

>>> satisfies @AsciiUpper Proxy 'A'
Nothing

>>> satisfies @AsciiUpper Proxy 'a'
Just (MkRefineException {predRep = AsciiUpper, targetRep = Char, msg = "a is not uppercase ascii"})

Since: 0.1.0.0

data AsciiLower

Predicate for isAsciiLower.

Examples

Expand

>>> satisfies @AsciiLower Proxy 'a'
Nothing

>>> satisfies @AsciiLower Proxy 'A'
Just (MkRefineException {predRep = AsciiLower, targetRep = Char, msg = "A is not lowercase ascii"})

Since: 0.1.0.0